Implemented AbstractRunAtInterval to seperate the 3 different periodic tasks in DonutClient. Added a kill method to the DonutPeer that kills the whole Donut program, which calls kill on the DonutClient and DonutServer objects.
Cleaned up code and formated with eclipse formatter.

Added constants to the thrift file for the intervals described above.

Modified the DonutJoinLeaveTest to take advantage of waiting depending on the current intervals.

When nodes go down due to no longer being alive, 3 places in code would print the stack trace with it. I commented these out since they seemed to be for debugging purposes only.

Moved the updatedSuccessorList function from the DonutClient into the Node file (since it directly operates on the node).